move image mode info variables to helper.py
