+++ Tests/helper.py
@@ -12,7 +12,7 @@ import sys
 import sysconfig
 import tempfile
 from io import BytesIO
-from typing import Any, Callable, Sequence
+from typing import Any, Callable, NamedTuple, Sequence
 
 import pytest
 from packaging.version import parse as parse_version
@@ -26,6 +26,40 @@ if os.environ.get("SHOW_ERRORS"):
     uploader = "show"
 elif "GITHUB_ACTIONS" in os.environ:
     uploader = "github_actions"
+
+
+class ImageModeInfo(NamedTuple):
+    name: str
+    pixel_size: int
+
+
+image_modes = (
+    ImageModeInfo("1", 1),
+    ImageModeInfo("L", 1),
+    ImageModeInfo("LA", 4),
+    ImageModeInfo("La", 4),
+    ImageModeInfo("P", 1),
+    ImageModeInfo("PA", 4),
+    ImageModeInfo("F", 4),
+    ImageModeInfo("I", 4),
+    ImageModeInfo("I;16", 2),
+    ImageModeInfo("I;16L", 2),
+    ImageModeInfo("I;16B", 2),
+    ImageModeInfo("I;16N", 2),
+    ImageModeInfo("RGB", 4),
+    ImageModeInfo("RGBA", 4),
+    ImageModeInfo("RGBa", 4),
+    ImageModeInfo("RGBX", 4),
+    ImageModeInfo("BGR;15", 2),
+    ImageModeInfo("BGR;16", 2),
+    ImageModeInfo("BGR;24", 3),
+    ImageModeInfo("CMYK", 4),
+    ImageModeInfo("YCbCr", 4),
+    ImageModeInfo("HSV", 4),
+    ImageModeInfo("LAB", 4),
+)
+
+image_mode_names = [mode.name for mode in image_modes]
 
 
 def upload(a: Image.Image, b: Image.Image) -> str | None:
